Reorganized packages
The packages were in a mess

They should be renamed to be clearer
