+++ src/main/java/yapper/commands/DeleteCommand.java
@@ -1,24 +1,24 @@
-package commands;
+package yapper.commands;
 
-import Yapper.Storage;
-import Yapper.Ui;
-import tasks.Task;
-import tasks.TaskList;
+import yapper.Storage;
+import yapper.Ui;
+import yapper.tasks.Task;
+import yapper.tasks.TaskList;
 
 import java.io.IOException;
 
-public class UnmarkCommand implements Command{
+public class DeleteCommand implements Command{
     private int index;
-    public UnmarkCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         Task task = tasks.get(index);
-        task.unmark();
-        ui.display("Ok, I've marked this task as not done yet");
-        ui.display(task.toString());
+        tasks.remove(index);
+        ui.display("Ok, I've removed this task");
+        ui.display(task.deleteMessage(tasks.size()));
         try {
             storage.save(tasks);
         } catch (IOException e) {
