+++ src/main/java/yapper/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package tasks;
+package yapper.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
