+++ src/main/java/yapper/tasks/Event.java
@@ -1,4 +1,4 @@
-package tasks;
+package yapper.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
