+++ src/main/java/yapper/tasks/Task.java
@@ -1,4 +1,4 @@
-package tasks;
+package yapper.tasks;
 
 import java.time.LocalDateTime;
